Changing the System to take to and from stations as input:
Modified the dataset: Removed the direction and replaced station by 
stationFrom and stationTo

Modified the TrainAgent: removed the dir instance member and replaced 
station member by stationFrom and stationTo. Made necessary modifications 
to achieve the same. Message that a TrainAgent has been created is now 
sent to both stations. When a message on the TrainAgent is displayed, 
the sender is now displayed as well.